# firmware/docs/openapi.yml
openapi: 3.1.0
info:
  title: Project Verdify MVP API
  version: "2.0"
  description: >
    REST API for AI-powered greenhouse MVP. HTTPS required for all endpoints. Metric units on wire.
    ETag used for config/plan. Device identity via device_name (verdify-aabbcc).
    Rate limiting applies to telemetry endpoints to prevent DoS attacks.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.verdify.ai/api/v1
...
security:
  - UserJWT: []
paths:
  /auth/register:
    post:
      tags: [Authentication]
      summary: Register new user account
      operationId: registerUser
      security: []
      parameters:
        - $ref: '#/components/parameters/CSRFTokenParam'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UserRegisterRequest' }
            examples:
              ex:
                value:
                  email: "user@example.com"
                  password: "secure_password"
                  full_name: "John Doe"
      responses:
        '201':
          description: User created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/UserRegisterResponse' }
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
  /auth/login:
    post:
      operationId: loginUser
      tags: [Authentication]
      summary: Login and obtain JWT
      security: []
      parameters:
        - $ref: '#/components/parameters/CSRFTokenParam'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UserLoginRequest' }
            examples:
              ex:
                value:
                  email: "user@example.com"
                  password: "secure_password"
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema: { $ref: '#/components/schemas/UserLoginResponse' }
        '401':
          $ref: '#/components/responses/Unauthorized'
  /auth/csrf:
    get:
      operationId: getCsrfToken
      tags: [Authentication]
      summary: Get CSRF token for browser-based authentication
      security: []
      responses:
        '200':
          description: CSRF token generated
          content:
            application/json:
              schema:
                type: object
                required: [csrf_token, expires_at]
                properties:
                  csrf_token:
                    type: string
                    description: CSRF token for use in X-CSRF-Token header
                  expires_at:
                    type: string
                    format: date-time
                    description: UTC timestamp when CSRF token expires
              examples:
                ex:
                  value:
                    csrf_token: "csrf_1234567890abcdef"
                    expires_at: "2025-08-13T19:05:00Z"
        '400':
          $ref: '#/components/responses/BadRequest'
...
  /telemetry/actuators:
    post:
      operationId: ingestActuatorTelemetry
      tags: [Telemetry]
      summary: Ingest actuator edge events
      security:
        - DeviceToken: []
      parameters:
        - $ref: '#/components/parameters/IdempotencyKeyParam'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/TelemetryActuators' }
      responses:
        '202':
          description: Accepted
          headers:
            X-RateLimit-Limit:
              description: Number of requests allowed per time window
              schema: { type: integer }
            X-RateLimit-Remaining:
              description: Number of requests remaining in current window
              schema: { type: integer }
            X-RateLimit-Reset:
              description: UTC timestamp when rate limit window resets
              schema: { type: integer }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/IngestResult' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
...
  /actuators:
    get:
      operationId: listActuators
      tags: [CRUD]
      summary: List actuators
      security:
        - UserJWT: []
      parameters:
        - name: controller_id
          in: query
          required: false
          schema: { type: string, format: uuid }
          description: Filter by controller ID
      responses:
        '200':
          description: List of actuators
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ActuatorList' }
        '401':
          $ref: '#/components/responses/Unauthorized'
    post:
      operationId: createActuator
      tags: [CRUD]
      summary: Create actuator
      security:
        - UserJWT: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ActuatorCreate' }
      responses:
        '201':
          description: Actuator created successfully
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Actuator' }
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /actuators/{id}:
    get:
      operationId: getActuator
      tags: [CRUD]
      summary: Get actuator details
      security:
        - UserJWT: []
      parameters:
        - $ref: '#/components/parameters/ActuatorIdParam'
      responses:
        '200':
          description: Actuator details
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Actuator' }
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      operationId: updateActuator
      tags: [CRUD]
      summary: Update actuator
      security:
        - UserJWT: []
      parameters:
        - $ref: '#/components/parameters/ActuatorIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ActuatorUpdate' }
      responses:
        '200':
          description: Actuator updated successfully
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Actuator' }
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      operationId: deleteActuator
      tags: [CRUD]
      summary: Delete actuator
      security:
        - UserJWT: []
      parameters:
        - $ref: '#/components/parameters/ActuatorIdParam'
      responses:
        '204': { description: Deleted }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
...
  /buttons:
    get:
      operationId: listButtons
      tags: [CRUD]
      summary: List controller buttons
      security:
        - UserJWT: []
      parameters:
        - name: controller_id
          in: query
          required: false
          schema: { type: string, format: uuid }
          description: Filter by controller ID
      responses:
        '200':
          description: List of controller buttons
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ControllerButtonList' }
        '401':
          $ref: '#/components/responses/Unauthorized'
    post:
      operationId: createButton
      tags: [CRUD]
      summary: Configure a controller button (manual override)
      security:
        - UserJWT: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ControllerButtonCreate' }
      responses:
        '201':
          description: Controller button configured successfully
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ControllerButton' }
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /buttons/{id}:
    get:
      operationId: getButton
      tags: [CRUD]
      summary: Get controller button details
      security:
        - UserJWT: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: Controller button details
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ControllerButton' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
    patch:
      operationId: updateButton
      tags: [CRUD]
      summary: Update controller button
      security:
        - UserJWT: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ControllerButtonUpdate' }
      responses:
        '200':
          description: Controller button updated successfully
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ControllerButton' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
    delete:
      operationId: deleteButton
      tags: [CRUD]
      summary: Delete controller button
      security:
        - UserJWT: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '204': { description: Deleted }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
...
components:
  securitySchemes:
    UserJWT:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: User JWT for app/admin calls
    DeviceToken:
      type: apiKey
      in: header
      name: X-Device-Token
      description: Long-lived device token for controllers
  parameters:
...
    ActuatorIdParam:
      in: path
      name: id
      required: true
      schema: { type: string, format: uuid }
...
    CSRFTokenParam:
      in: header
      name: X-CSRF-Token
      required: false
      description: CSRF token for browser-based requests (optional for non-browser clients)
      schema: { type: string }
    IdempotencyKeyParam:
      in: header
      name: Idempotency-Key
      required: false
      description: Idempotency key for exactly-once processing (recommended for telemetry)
      schema: { type: string, maxLength: 255 }
  responses:
    BadRequest:
      description: Bad request
      headers:
        X-Request-Id:
          description: Request identifier for tracing
          schema: { type: string }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
          examples:
            ex:
              value:
                error_code: E400_BAD_REQUEST
                message: "Invalid input"
                timestamp: "2025-08-13T18:05:00Z"
                details:
                  field: "email"
                  reason: "Invalid email format"
                request_id: "req_1234567890abcdef"
    Unauthorized:
      description: Unauthorized
      headers:
        X-Request-Id:
          description: Request identifier for tracing
          schema: { type: string }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
          examples:
            ex:
              value:
                error_code: E401_UNAUTHORIZED
                message: "Missing/invalid token"
                timestamp: "2025-08-13T18:05:00Z"
                request_id: "req_1234567890abcdef"
    NotFound:
      description: Not found
      headers:
        X-Request-Id:
          description: Request identifier for tracing
          schema: { type: string }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
          examples:
            ex:
              value:
                error_code: E404_NOT_FOUND
                message: "Resource not found"
                timestamp: "2025-08-13T18:05:00Z"
                request_id: "req_1234567890"
    Conflict:
      description: Conflict (uniqueness violated)
      headers:
        X-Request-Id:
          description: Request identifier for tracing
          schema: { type: string }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
          examples:
            ex:
              value:
                error_code: E409_CONFLICT
                message: "Unique constraint violated"
                timestamp: "2025-08-13T18:05:00Z"
                request_id: "req_1234567890"
    TooManyRequests:
      description: Rate limit exceeded
      headers:
        X-RateLimit-Limit:
          description: Number of requests allowed per time window
          schema: { type: integer }
        X-RateLimit-Remaining:
          description: Number of requests remaining in current window
          schema: { type: integer }
        X-RateLimit-Reset:
          description: Epoch seconds when the window resets
          schema: { type: integer }
        Retry-After:
          description: Seconds to wait before retry
          schema: { type: integer }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
          examples:
            ex:
              value:
                error_code: E429_TOO_MANY_REQUESTS
                message: "Rate limit exceeded. Try again in 60 seconds."
                timestamp: "2025-08-13T18:05:00Z"
                request_id: "req_1234567890abcdef"
...
    UnprocessableEntity:
      description: Unprocessable entity - validation failed
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
          examples:
            ex:
              value:
                error_code: E422_UNPROCESSABLE
                message: "Validation failed"
                timestamp: "2025-08-13T18:05:00Z"
                details:
                  - field: "email"
                    value: "invalid-email"
                    expected: "Valid email format"
                request_id: "req_1234567890abcdef"
  schemas:
...
    ActuatorKind:
      type: string
      enum: [fan, heater, vent, fogger, irrigation_valve, fertilizer_valve, pump, light]
    FailSafeState:
      type: string
      enum: [on, off]
    ButtonKind:
      type: string
      enum: [cool, heat, humid]

    # --- Error response ---
    ErrorResponse:
      type: object
      required: [error_code, message, timestamp, request_id]
      properties:
        error_code:
          $ref: '#/components/schemas/ErrorCode'
          description: Standardized error code
          example: "E409_CONFLICT"
        message:
          type: string
          description: Human-readable error message
          example: "Email already exists"
        timestamp:
          type: string
          format: date-time
          description: ISO 8601 timestamp of when the error occurred
        request_id:
          type: string
          description: Unique request identifier for tracing
          example: "req_1234567890"
        details:
          oneOf:
            - type: array
              items: { $ref: '#/components/schemas/ErrorDetail' }
            - type: object
              additionalProperties: true
            - type: string
          description: Structured error details for field-specific issues
          example:
            field: "email"
            value: "user@example.com"
...
    UserRegisterRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
        full_name:
          type: [string, "null"]
    UserRegisterResponse:
      type: object
      required: [user, access_token]
      properties:
        user:
          type: object
          required: [id, email, created_at]
          properties:
            id: { type: string, format: uuid }
            email: { type: string, format: email }
            full_name: { type: [string, "null"] }
            created_at: { type: string, format: date-time }
        access_token:
          type: string
    UserLoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
    UserLoginResponse:
      type: object
      required: [access_token, token_type, expires_in]
      properties:
        access_token:
          type: string
        token_type:
          type: string
          enum: [bearer]
          default: bearer
        expires_in:
          type: integer
          description: Token expiry in seconds

    # --- Onboarding ---
...
    Paginated:
      type: object
      required: [data, page, page_size, total]
      properties:
        data:
          type: array
          description: Array of items for current page
        page:
          type: integer
          minimum: 1
          description: Current page number
        page_size:
          type: integer
          minimum: 1
          maximum: 200
          description: Number of items per page
        total:
          type: integer
          minimum: 0
          description: Total number of items across all pages

    # --- CRUD core entities ---
...
    Actuator:
      type: object
      required: [id, controller_id, name, kind]
      properties:
        id: { type: string, format: uuid }
        controller_id: { type: string, format: uuid }
        name: { type: string }
        kind: { $ref: '#/components/schemas/ActuatorKind' }
        relay_channel: { type: [integer, "null"] }
        min_on_ms: { type: integer, default: 60000 }
        min_off_ms: { type: integer, default: 60000 }
        fail_safe_state: { $ref: '#/components/schemas/FailSafeState' }
        zone_id: { type: [string, "null"], format: uuid }
    ActuatorCreate:
      type: object
      required: [controller_id, name, kind]
      properties:
        controller_id: { type: string, format: uuid }
        name: { type: string }
        kind: { $ref: '#/components/schemas/ActuatorKind' }
        relay_channel: { type: [integer, "null"] }
        min_on_ms: { type: integer, default: 60000 }
        min_off_ms: { type: integer, default: 60000 }
        fail_safe_state: { $ref: '#/components/schemas/FailSafeState' }
        zone_id: { type: [string, "null"], format: uuid }
    ActuatorUpdate:
      type: object
      properties:
        name: { type: string }
        kind: { $ref: '#/components/schemas/ActuatorKind' }
        relay_channel: { type: [integer, "null"] }
        min_on_ms: { type: integer }
        min_off_ms: { type: integer }
        fail_safe_state: { $ref: '#/components/schemas/FailSafeState' }
        zone_id: { type: [string, "null"], format: uuid }
    ActuatorList:
      allOf:
        - $ref: '#/components/schemas/Paginated'
        - type: object
          properties:
            data:
              type: array
              items: { $ref: '#/components/schemas/Actuator' }
...
    ControllerButton:
      type: object
      required: [id, controller_id, button_kind, timeout_s]
      properties:
        id: { type: string, format: uuid }
        controller_id: { type: string, format: uuid }
        button_kind: { $ref: '#/components/schemas/ButtonKind' }
        target_temp_stage: { anyOf: [{ type: integer, minimum: -3, maximum: 3 }, { type: "null" }] }
        target_humi_stage: { anyOf: [{ type: integer, minimum: -3, maximum: 3 }, { type: "null" }] }
        timeout_s: { type: integer, minimum: 1 }
    ControllerButtonCreate:
      type: object
      required: [controller_id, button_kind, timeout_s]
      properties:
        controller_id: { type: string, format: uuid }
        button_kind: { $ref: '#/components/schemas/ButtonKind' }
        target_temp_stage: { anyOf: [{ type: integer, minimum: -3, maximum: 3 }, { type: "null" }] }
        target_humi_stage: { anyOf: [{ type: integer, minimum: -3, maximum: 3 }, { type: "null" }] }
        timeout_s: { type: integer, minimum: 1 }
    ControllerButtonUpdate:
      type: object
      properties:
        button_kind: { $ref: '#/components/schemas/ButtonKind' }
        target_temp_stage: { anyOf: [{ type: integer, minimum: -3, maximum: 3 }, { type: "null" }] }
        target_humi_stage: { anyOf: [{ type: integer, minimum: -3, maximum: 3 }, { type: "null" }] }
        timeout_s: { type: integer, minimum: 1 }
    ControllerButtonList:
      allOf:
        - $ref: '#/components/schemas/Paginated'
        - type: object
          properties:
            data:
              type: array
              items: { $ref: '#/components/schemas/ControllerButton' }
...
    TelemetryActuators:
      type: object
      required: [events]
      properties:
        events:
          type: array
          items:
            type: object
            required: [actuator_id, ts_utc, state, reason]
            properties:
              actuator_id: { type: string, format: uuid }
              ts_utc: { type: string, format: date-time }
              state: { type: boolean }
              reason: { type: string }
...
    IngestResult:
      type: object
      properties:
        accepted: { type: integer }
        rejected: { type: integer }
        errors:
          type: array
          items: { $ref: '#/components/schemas/ErrorResponse' }